Move map output helpers into partial
